+++ src/lib/skilling/skills/smithing/smelting.ts
@@ -1,6 +1,6 @@
-import itemID from '../../util/itemID';
-import { Emoji } from '../../constants';
-import { Bar, SkillsEnum } from '../types';
+import itemID from '../../../util/itemID';
+import { Emoji } from '../../../constants';
+import { Bar, SkillsEnum } from '../../types';
 
 const bars: Bar[] = [
 	{
